Backend/requirements.txt
adafruit-io==2.5.0
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
paho-mqtt==1.5.1
Pillow==8.2.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
zope.interface==5.4.0
